requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
defusedxml==0.7.1
dj-database-url==2.1.0
dj-email-url==1.0.6
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.57.0
django-cache-url==3.4.5
django-debug-toolbar==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
environs==10.0.0
flower==2.0.1
httpie==3.2.2
humanize==4.9.0
idna==3.4
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.4
markdown-it-py==3.0.0
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.6.0
rpds-py==0.15.2
six==1.16.0
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
